requirements.txt
torch
youtube-transcript-api
Flask
groq
langchain
langchain-core
langchain-groq
langchain-text-splitters
langsmith
python-dotenv
numpy
transformers 
torch
